fixes it? Adds factories that are just standard readables in _1,_2, adds a factory of the actual device to _optics, makes dcm.py a standard readable with energy and wavelength
