+++ src/dodal/beamlines/i19_1.py
@@ -13,8 +13,7 @@ from dodal.devices.beamlines.i19.access_controlled.attenuator_motor_squad import
     AttenuatorMotorSquad,
 )
 from dodal.devices.beamlines.i19.access_controlled.blueapi_device import HutchState
-
-# from dodal.devices.beamlines.i19.access_controlled.dcm import DCM
+from dodal.devices.beamlines.i19.access_controlled.dcm import DCM
 from dodal.devices.beamlines.i19.access_controlled.piezo_control import (
     AccessControlledPiezoActuator,
     FocusingMirrorName,
@@ -69,9 +68,9 @@ def config_client() -> ConfigClient:
     return client
 
 
-# @devices.factory()
-# def dcm() -> DCM:
-#     return DCM(f"{PREFIX.beamline_prefix}-MO-DCM-01:") ?
+@devices.factory()
+def dcm() -> DCM:
+    return DCM(f"{PREFIX.beamline_prefix}-MO-DCM-01:")
 
 
 @devices.factory()
