+++ src/dodal/beamlines/i19_2.py
@@ -14,8 +14,7 @@ from dodal.devices.beamlines.i19.access_controlled.attenuator_motor_squad import
     AttenuatorMotorSquad,
 )
 from dodal.devices.beamlines.i19.access_controlled.blueapi_device import HutchState
-
-# from dodal.devices.beamlines.i19.access_controlled.dcm import DCM
+from dodal.devices.beamlines.i19.access_controlled.dcm import DCM
 from dodal.devices.beamlines.i19.access_controlled.piezo_control import (
     AccessControlledPiezoActuator,
     FocusingMirrorName,
@@ -65,9 +64,9 @@ def path_provider() -> PathProvider:
     )
 
 
-# @devices.factory()
-# def dcm() -> DCM:
-#     return DCM(f"{PREFIX.beamline_prefix}-MO-DCM-01:") ?
+@devices.factory()
+def dcm() -> DCM:
+    return DCM(f"{PREFIX.beamline_prefix}-MO-DCM-01:")
 
 
 @devices.factory()
